first version
the initial version of dump output to the command line has been implemented, blanks of functions for dump output to the log have been made, and adding elements to the tree has also been implemented. It is necessary to implement dump with a system call, as well as functions for deleting elements.