added delete and dump function
Added a function to remove an element from a tree by its value, the function can only remove leaves, otherwise the function returns an error, nothing else happens. Also, to find the element to be removed, a function for recursive tree traversal has been implemented, and a variable pointer to the parent has been added to the root structure. Also, a basic graphic dump has been implemented, with the tree output to a file.